# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /limits/workspaces/{workspaceId}:
    get:
      operationId: getSetLimitsWorkspace
      tags: [limits]
      summary: Get the composed Set Limits read model
      description: Reads persisted directory, billing-cycle usage, and limit observations without refreshing upstream data.
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdPath"
      responses:
        "200":
          description: Set Limits workspace model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetLimitsWorkspace"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Workspace not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /limits/operations/prepare:
    post:
      operationId: prepareLimitOperation
      tags: [limits]
      summary: Freeze and review workspace member-limit targets
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LimitOperationPrepareInput" }
      responses:
        "200":
          description: Frozen operation review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LimitOperation" }
        "400":
          description: Invalid amount or target selection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: Idempotency key was reused for a different request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /limits/operations/{operationId}/commit:
    post:
      operationId: commitLimitOperation
      tags: [limits]
      summary: Commit an exactly reviewed limit operation
      parameters:
        - $ref: "#/components/parameters/LimitOperationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LimitOperationCommitInput" }
      responses:
        "202":
          description: Durable operation accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LimitOperation" }
        "400":
          description: Commit does not match the frozen review
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Operation not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "409":
          description: A selected member already has an active operation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /limits/operations/{operationId}:
    get:
      operationId: getLimitOperation
      tags: [limits]
      summary: Get and resume a durable limit operation
      parameters:
        - $ref: "#/components/parameters/LimitOperationIdPath"
      responses:
        "200":
          description: Current operation and per-target status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LimitOperation" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Operation not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
  /limits/operations/{operationId}/retry:
    post:
      operationId: retryLimitOperationTargets
      tags: [limits]
      summary: Retry selected failed or verification-pending targets
      parameters:
        - $ref: "#/components/parameters/LimitOperationIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LimitOperationRetryInput" }
      responses:
        "202":
          description: Eligible targets queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LimitOperation" }
        "400":
          description: No eligible targets selected
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: Operation not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
components:
  responses:
...
    Unauthorized:
      description: Authentication required — no valid session.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Forbidden:
      description: >
        Authenticated but not authorized (neither account admin nor an enabled
        workspace admin, or attempting an account-admin-only action).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  parameters:
    LimitOperationIdPath:
      name: operationId
      in: path
      required: true
      schema:
        type: string
        pattern: '^[0-9a-fA-F-]{36}$'
    WorkspaceIdPath:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  schemas:
    SetLimitsGroup:
      type: object
      required: [groupId, name, familyName, role, eligibleUserIds]
      properties:
        groupId: { type: string }
        name: { type: string }
        familyName: { type: ["string", "null"] }
        role: { type: ["string", "null"] }
        eligibleUserIds:
          type: array
          items: { type: string, pattern: '^[1-9]\d*$' }
    SetLimitsMember:
      type: object
      required: [userId, username, name, email, role, groupIds, isInternal, isDisabled, eligible, usageUsd, explicitLimitUsd, effectiveLimitUsd, limitState]
      properties:
        userId: { type: string, pattern: '^[1-9]\d*$' }
        username: { type: string }
        name: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        role: { type: string }
        groupIds:
          type: array
          items: { type: string }
        isInternal: { type: boolean }
        isDisabled: { type: boolean }
        eligible: { type: boolean }
        usageUsd: { type: ["number", "null"] }
        explicitLimitUsd: { type: ["number", "null"] }
        effectiveLimitUsd: { type: ["number", "null"] }
        limitState:
          type: string
          enum: [explicit, inherited, no_limit, unavailable]
    SetLimitsWorkspace:
      type: object
      required: [workspaceId, workspaceName, canWrite, unavailableReason, billingPeriod, limitObservation, groups, members]
      properties:
        workspaceId: { type: string, pattern: '^[A-Za-z0-9]+$' }
        workspaceName: { type: string }
        canWrite: { type: boolean }
        unavailableReason: { type: ["string", "null"] }
        billingPeriod:
          type: object
          required: [start, end]
          properties:
            start: { type: string, format: date-time }
            end: { type: string, format: date-time }
        limitObservation:
          type: object
          required: [status, observedAt, error]
          properties:
            status: { type: string, enum: [available, unavailable, failed] }
            observedAt: { type: ["string", "null"], format: date-time }
            error: { type: ["string", "null"] }
        groups:
          type: array
          items: { $ref: "#/components/schemas/SetLimitsGroup" }
        members:
          type: array
          items: { $ref: "#/components/schemas/SetLimitsMember" }
    LimitOperationPrepareInput:
      type: object
      additionalProperties: false
      required: [workspaceId, amountUsd, userIds, groupIds, idempotencyKey]
      properties:
        workspaceId: { type: string, pattern: '^[A-Za-z0-9]+$' }
        amountUsd: { type: number, exclusiveMinimum: 0, multipleOf: 0.01 }
        userIds:
          type: array
          maxItems: 1000
          items: { type: string, pattern: '^[1-9]\d*$' }
        groupIds:
          type: array
          maxItems: 1000
          items: { type: string, minLength: 1 }
        idempotencyKey: { type: string, minLength: 8, maxLength: 200 }
    LimitOperationCommitInput:
      type: object
      additionalProperties: false
      required: [reviewFingerprint, amountUsd, userIds]
      properties:
        reviewFingerprint: { type: string, minLength: 64, maxLength: 64 }
        amountUsd: { type: number, exclusiveMinimum: 0, multipleOf: 0.01 }
        userIds:
          type: array
          minItems: 1
          maxItems: 1000
          items: { type: string, pattern: '^[1-9]\d*$' }
    LimitOperationRetryInput:
      type: object
      additionalProperties: false
      required: [userIds, idempotencyKey]
      properties:
        userIds:
          type: array
          minItems: 1
          maxItems: 1000
          items: { type: string, pattern: '^[1-9]\d*$' }
        idempotencyKey: { type: string, minLength: 8, maxLength: 200 }
    LimitTargetAttempt:
      type: object
      required: [at, stage, outcome]
      properties:
        at: { type: string, format: date-time }
        stage: { type: string, enum: [authorization, membership, reconcile, write, verification, audit] }
        outcome: { type: string }
        requestId: { type: ["string", "null"] }
        retryAfterMs: { type: ["integer", "null"] }
        message: { type: ["string", "null"] }
    LimitOperationTarget:
      type: object
      required: [workspaceId, userId, memberName, memberEmail, oldAmountUsd, newAmountUsd, state, attempts, history, errorStage, errorCode, errorMessage, upstreamRequestId, queuedAt, applyingAt, verifiedAt, failedAt]
      properties:
        workspaceId: { type: string }
        userId: { type: string }
        memberName: { type: ["string", "null"] }
        memberEmail: { type: ["string", "null"] }
        oldAmountUsd: { type: ["number", "null"] }
        newAmountUsd: { type: number }
        state: { type: string, enum: [queued, applying, verified, failed, verification_pending] }
        attempts: { type: integer }
        history:
          type: array
          items: { $ref: "#/components/schemas/LimitTargetAttempt" }
        errorStage: { type: ["string", "null"] }
        errorCode: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        upstreamRequestId: { type: ["string", "null"] }
        queuedAt: { type: ["string", "null"], format: date-time }
        applyingAt: { type: ["string", "null"], format: date-time }
        verifiedAt: { type: ["string", "null"], format: date-time }
        failedAt: { type: ["string", "null"], format: date-time }
    LimitOperation:
      type: object
      required: [id, workspaceId, state, amountUsd, reviewFingerprint, actorUserId, preparedAt, committedAt, completedAt, counts, targets]
      properties:
        id: { type: string, pattern: '^[0-9a-fA-F-]{36}$' }
        workspaceId: { type: string }
        state: { type: string, enum: [prepared, queued, running, completed] }
        amountUsd: { type: number }
        reviewFingerprint: { type: string }
        actorUserId: { type: string }
        preparedAt: { type: string, format: date-time }
        committedAt: { type: ["string", "null"], format: date-time }
        completedAt: { type: ["string", "null"], format: date-time }
        counts:
          type: object
          required: [total, queued, applying, verified, failed, verificationPending]
          properties:
            total: { type: integer }
            queued: { type: integer }
            applying: { type: integer }
            verified: { type: integer }
            failed: { type: integer }
            verificationPending: { type: integer }
        targets:
          type: array
          items: { $ref: "#/components/schemas/LimitOperationTarget" }
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...