# lib/api-zod/src/generated/types/limitOperation.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LimitOperationCounts } from './limitOperationCounts';
import type { LimitOperationState } from './limitOperationState';
import type { LimitOperationTarget } from './limitOperationTarget';

export interface LimitOperation {
  /** @pattern ^[0-9a-fA-F-]{36}$ */
  id: string;
  workspaceId: string;
  state: LimitOperationState;
  amountUsd: number;
  reviewFingerprint: string;
  actorUserId: string;
  preparedAt: Date;
  /** @nullable */
  committedAt: Date | null;
  /** @nullable */
  completedAt: Date | null;
  counts: LimitOperationCounts;
  targets: LimitOperationTarget[];
}