# lib/api-zod/src/generated/types/setLimitsMember.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SetLimitsMemberLimitState } from './setLimitsMemberLimitState';

export interface SetLimitsMember {
  /** @pattern ^[1-9]\d*$ */
  userId: string;
  username: string;
  /** @nullable */
  name: string | null;
  /** @nullable */
  email: string | null;
  role: string;
  groupIds: string[];
  isInternal: boolean;
  isDisabled: boolean;
  eligible: boolean;
  /** @nullable */
  usageUsd: number | null;
  /** @nullable */
  explicitLimitUsd: number | null;
  /** @nullable */
  effectiveLimitUsd: number | null;
  limitState: SetLimitsMemberLimitState;
}