# lib/api-zod/src/generated/types/setLimitsWorkspace.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SetLimitsGroup } from './setLimitsGroup';
import type { SetLimitsMember } from './setLimitsMember';
import type { SetLimitsWorkspaceBillingPeriod } from './setLimitsWorkspaceBillingPeriod';
import type { SetLimitsWorkspaceLimitObservation } from './setLimitsWorkspaceLimitObservation';

export interface SetLimitsWorkspace {
  /** @pattern ^[A-Za-z0-9]+$ */
  workspaceId: string;
  workspaceName: string;
  canWrite: boolean;
  /** @nullable */
  unavailableReason: string | null;
  billingPeriod: SetLimitsWorkspaceBillingPeriod;
  limitObservation: SetLimitsWorkspaceLimitObservation;
  groups: SetLimitsGroup[];
  members: SetLimitsMember[];
}